requirements.txt
asgiref==3.6.0
attrs==23.1.0
beautifulsoup4==4.12.2
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
inflection==0.5.1
jsonschema==4.17.3
psycopg2==2.9.6
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
soupsieve==2.4.1
sqlparse==0.4.4
uritemplate==4.1.1
